Revisions
Add fetch_property to _action/_utils.py and resolve circular import
