+++ src/stratis_cli/_actions/_top.py
@@ -51,25 +51,7 @@ def _fetch_keylist_property(proxy):
     :raises StratisCliPropertyNotFoundError:
     :raises StratisCliEnginePropertyError:
     """
-    return _fetch_property(proxy, "KeyList")
-
-
-def _fetch_property(proxy, property_name):
-    """
-    Fetch a property from stratisd.
-    :param proxy: proxy to the top object in stratisd
-    :param str property_name: name of the property to fetch
-    :return: value associated with the requested property name
-    :raises StratisCliPropertyNotFoundError:
-    :raises StratisCliEnginePropertyError:
-    """
-    # pylint: disable=import-outside-toplevel
-    from ._data import FetchProperties
-
-    properties = FetchProperties.Methods.GetProperties(
-        proxy, {"properties": [property_name]}
-    )
-    return fetch_property(properties, property_name)
+    return fetch_property(proxy, "KeyList")
 
 
 def _add_update_key(proxy, key_desc, capture_key, *, keyfile_path):
